Feature Update
Update utils and style components.
