+++ src/Components/Header/Header.js
@@ -1,13 +1,8 @@
 import React from "react";
 import HeaderDate from "./HeaderDate";
 import HeaderDay from "./HeaderDay";
-import styled from "styled-components";
+import HeaderStyle from "../../styled/headerStyle";
 
-const HeaderStyle = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-`;
 const Header = () => {
   return (
     <>
