+++ src/Components/Main-Section/Main.js
@@ -1,8 +1,7 @@
 import Footer from "../Footer/Footer";
 import Header from "../Header/Header";
-import ToDoList from "../ToDo-Container/ToDoList";
-import styled from "styled-components";
-
+import ToDoList from "../ToDo-Section/ToDoList";
+import MainStyle from "../../styled/mainStyle";
 
 const ToDoData = [
 
@@ -12,18 +11,12 @@ const ToDoData = [
   {id: 4, data : "Grab some Pizza"},
   {id: 5, data : "Grab some Pizza"},
   {id: 6, data : "Grab some Pizza"},
+  {id: 7, data : "Grab some Pizza"},
+  {id: 8, data : "Grab some Pizza"},
+
  ];
  
-const MainStyle = styled.div`
-    background: #F0EFE9;
-    position: relative ;
-    background: whitesmoke;
-    width: 260px;
-    margin: 10em auto;
-    padding: 20px 10px;
-    color: black;
-    box-shadow: -1px -1px 6px 3px rgba(0,0,0,0.1);
-  `
+
 
 function Main() {
   return (
