--- src/Components/ToDo-Container/ToDoTask.css
@@ -1,7 +1,0 @@
-.complete{
-    color: green;
-}
-
-.not-complete{
-    color: grey;
-}
