+++ src/styled/Date/dateStyle.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+const DateStyle = styled.div`
+  margin: 0px 2px 0px 0px;
+  padding: 0px;
+  font-size: 2.4rem;
+  color: black;
+  font-weight: 400;
+`;
+export default DateStyle;
