+++ src/styled/Date/headerDateStyle.js
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+const HeaderDateStyle = styled.div`
+  display: flex;
+  margin: 0.5em;
+  text-align: center;
+`;
+
+export default HeaderDateStyle;
