+++ src/styled/Date/headerMonthYear.js
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+const MonthYear = styled.div`
+  display: flex;
+  flex-direction: column;
+  padding: 4px;
+  margin: 0px;
+`;
+
+export default MonthYear
