+++ src/styled/Date/monthStyle.js
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+const MonthStyle = styled.div`
+  color: black;
+  font-weight: 400;
+  font-size: 1rem;
+`;
+export default MonthStyle;
