+++ src/styled/Date/yearStyle.js
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+const YearStyle = styled.div`
+  color: grey;
+  font-weight: 400;
+  font-size: 1rem;
+`;
+
+export default YearStyle;
