+++ src/styled/Day/dayStyle.js
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+const HeaderDayStyle = styled.div`
+display: flex;
+justify-content: center;
+align-items: center;
+  margin: 0.5em;
+  font-size: 1em;
+  text-align: center;
+  color: black;
+  font-weight: 400;
+`;
+export default HeaderDayStyle
