+++ src/styled/ToDo/listStyle.js
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+const ToDoListStyle = styled.div`
+  display: flex;
+  flex-direction: column;
+  height: 280px;
+  overflow-y: auto;
+  -ms-overflow-style: none;
+  scrollbar-width: none;
+`;
+export default ToDoListStyle;
