+++ src/styled/ToDo/taskStyle.js
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+const ToDOTaskStyle = styled.div`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 3px;
+  margin: 0.5em;
+`;
+export default ToDOTaskStyle;
