+++ src/styled/headerStyle.js
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+const HeaderStyle = styled.div`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+`;
+
+export default HeaderStyle
