+++ src/styled/mainStyle.js
@@ -1,0 +1,15 @@
+import styled from "styled-components";
+
+const MainStyle = styled.div`
+  background: #f0efe9;
+  position: relative;
+  background: whitesmoke;
+  width: 260px;
+  height: 360px;
+  margin: 10em auto;
+  padding: 20px 10px;
+  color: black;
+  box-shadow: -1px -1px 6px 3px rgba(0, 0, 0, 0.1);
+`;
+
+export default MainStyle;
